requirements.txt
# Core ML Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.11.0

# Data Processing
pyarrow>=6.0.0
geopandas>=0.10.0
shapely>=1.8.0

# MLOps & Experiment Tracking
mlflow>=1.20.0
prefect>=2.0.0
wandb>=0.12.0

# API & Serving
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Testing & Development
pytest>=6.2.5
pytest-cov>=2.12.0
black>=21.9b0
flake8>=3.9.0
mypy>=0.910

# Configuration & Logging
pyyaml>=5.4.1
python-dotenv>=0.19.0
structlog>=21.1.0

# Monitoring & Metrics
prometheus-client>=0.11.0
grafana-api>=1.0.3

# Cloud & Container
docker>=5.0.0
boto3>=1.18.0
google-cloud-storage>=1.42.0

# Visualization
matplotlib>=3.4.3
seaborn>=0.11.2
plotly>=5.3.1
